+++ src/transaction.local.net.spec.ts
@@ -1,7 +1,6 @@
 import { Transaction } from "./transaction";
 import { GasLimit } from "./networkParams";
 import { TransactionPayload } from "./transactionPayload";
-import { NetworkConfig } from "./networkConfig";
 import { Balance } from "./balance";
 import { loadTestWallets, TestWallet } from "./testutils";
 import { Logger } from "./logger";
@@ -11,6 +10,7 @@ import { TransactionWatcher } from "./transactionWatcher";
 
 describe("test transaction", function () {
     let alice: TestWallet, bob: TestWallet;
+
     before(async function () {
         ({ alice, bob } = await loadTestWallets());
     });
@@ -20,8 +20,8 @@ describe("test transaction", function () {
 
         let provider = chooseProxyProvider("local-testnet");
         let watcher = new TransactionWatcher(provider);
-
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
+        
         await alice.sync(provider);
 
         await bob.sync(provider);
@@ -29,12 +29,16 @@ describe("test transaction", function () {
 
         let transactionOne = new Transaction({
             receiver: bob.address,
-            value: Balance.egld(42)
+            value: Balance.egld(42),
+            gasLimit: network.MinGasLimit,
+            chainID: network.ChainID
         });
 
         let transactionTwo = new Transaction({
             receiver: bob.address,
-            value: Balance.egld(43)
+            value: Balance.egld(43),
+            gasLimit: network.MinGasLimit,
+            chainID: network.ChainID
         });
 
         transactionOne.setNonce(alice.account.nonce);
@@ -60,22 +64,23 @@ describe("test transaction", function () {
         this.timeout(20000);
 
         let provider = chooseProxyProvider("local-testnet");
-
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         let transactionOne = new Transaction({
             data: new TransactionPayload("helloWorld"),
             gasLimit: new GasLimit(70000),
             receiver: alice.address,
-            value: Balance.egld(1000)
+            value: Balance.egld(1000),
+            chainID: network.ChainID
         });
 
         let transactionTwo = new Transaction({
             data: new TransactionPayload("helloWorld"),
             gasLimit: new GasLimit(70000),
             receiver: alice.address,
-            value: Balance.egld(1000000)
+            value: Balance.egld(1000000),
+            chainID: network.ChainID
         });
 
         transactionOne.setNonce(alice.account.nonce);